requirements.txt
fastapi
uvicorn
pytesseract
boto3
pillow
langchain
langchain-groq
python-dotenv
